added processes manager config
