+++ orchestrator/config/config_files/processes_manager.yaml
@@ -1,0 +1,4 @@
+process_groups:
+  - name: "executor"
+    replicas: 3
+    command: "./start_executor.sh"
